+++ src/index.js
@@ -12,6 +12,7 @@ import './js/darkMode.js';
 import './js/to-top.js';
 import './js/buttonEffects.js';
 import './js/showFilms.js';
+import './js/searchFormListener.js';
 
 import { initializeApp } from 'firebase/app';
 import { getAuth } from 'firebase/auth';
