+++ src/js/searchFormListener.js
@@ -1,0 +1,20 @@
+const searchForm = document.querySelector('#search-form');
+
+searchForm.addEventListener('submit', evt => {
+  evt.preventDefault();
+  if (searchForm.elements.searchQuery.value.trim() == '') {
+    Notiflix.Notify.warning('Enter some text...');
+  } else {
+    fetchFilmsByQuery(searchForm.elements.searchQuery.value.trim().split(' ').join(`%20`), 1)
+      .then(res => {
+        if (res.data.total_results == 0) {
+          errorText.classList.remove('hiddenVisibility');
+        } else {
+          errorText.classList.add('hiddenVisibility');
+          Notiflix.Notify.success(`Znaeziono: ${res.data.total_results}`);
+          showFilms(res);
+        }
+      })
+      .catch(error => {});
+  }
+});
